Security_v2

+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserServiceImpl.java
@@ -1,6 +1,8 @@
 package ru.kata.spring.boot_security.demo.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.kata.spring.boot_security.demo.repository.UserRepository;
@@ -8,13 +10,21 @@ import ru.kata.spring.boot_security.demo.exception.UserNotFoundException;
 import ru.kata.spring.boot_security.demo.entity.User;
 
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
+@NoArgsConstructor
+@AllArgsConstructor
 @Transactional
 public class UserServiceImpl implements UserService {
 
-    @Autowired
+
     private UserRepository userRepository;
+
+    private RoleService roleService;
+
+    private PasswordEncoder passwordEncoder;
 
     @Override
     public List<User> getAllUsers() {
@@ -22,35 +32,51 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    @Transactional(noRollbackFor = UserNotFoundException.class)
     public User findById(Long id) throws UserNotFoundException {
         return userRepository.findById(id).orElseThrow(() -> new UserNotFoundException("couldn't find user with id " + id));
     }
 
     @Override
+    @Transactional
     public void save(User user) {
-
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
         userRepository.save(user);
+
     }
 
     @Override
+    @Transactional
     public void saveAll(List<User> users) {
         userRepository.saveAll(users);
     }
     @Override
-    public void update(User updatedUser) {
+    @Transactional
+    public void update(User updatedUser, Set<Long> roleIds) {
+        User user = findById(updatedUser.getId());
+
+        if(roleIds != null) {
+            user.setRoles(roleIds.stream()
+                    .map(roleService::findById)
+                    .collect(Collectors.toSet()));
+        }
+        user.setUsername(updatedUser.getUsername());
+        user.setEmail(updatedUser.getEmail());
+        user.setPhone(updatedUser.getPhone());
+        if(updatedUser.getPassword() != null && !updatedUser.getPassword().isEmpty()) {
+            updatedUser.setPassword(passwordEncoder.encode(updatedUser.getPassword()));
+        }
         userRepository.save(updatedUser);
     }
 
     @Override
+    @Transactional
     public void delete(Long id) {
-        if (id > 0) {
+        if (id > 0 && id != null) {
             userRepository.deleteById(id);
         }
     }
 
     @Override
-    @Transactional(noRollbackFor = UserNotFoundException.class)
     public User findByUsername(String username) throws UserNotFoundException {
         return userRepository.getUserByUsername(username).orElseThrow(() -> new UserNotFoundException("couldn't find user with username " +username));
     }
